% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVariable.R
\name{as_data_dictionary}
\alias{as_data_dictionary}
\title{Create Data Dictionary from Existing Data}
\usage{
as_data_dictionary(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a \code{DataDictionary}
}
\description{
Create Data Dictionary from Existing Data
}
