% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVariable.R
\name{as_tibble.DataDictionary}
\alias{as_tibble.DataDictionary}
\title{Coerce data dictionary to data frame}
\usage{
as_tibble.DataDictionary(x, ...)
}
\arguments{
\item{x}{a \code{DataDictionary} object}

\item{...}{}
}
\value{
a \code{tibble}
}
\description{
Coerce data dictionary to data frame
}
\examples{

age_years <- NumericVariable$new(
  name = "age",
  label = "Age of participant"
)

age_group <- NominalVariable$new(
  name = "age_group",
  label = "Age group",
  category_levels = c("age_lt_50", "age_gteq_50_lt_60", "age_gteq_60"),
  category_labels = c("0 to < 50", "50 to < 60", "\u2265 60")
)

data_dictionary(age_years, age_group)
}
