% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVariable.R
\name{data_dictionary}
\alias{data_dictionary}
\title{Create Data Dictionary from Variable Definitions}
\usage{
data_dictionary(...)
}
\arguments{
\item{...}{One or more objects inheriting from \code{NumericVariable}
or \code{NominalVariable}.}
}
\value{
A \code{DataDictionary} object containing a tibble summary of all
variables.
}
\description{
Initializes a \code{DataDictionary} object from a set of variables.
Variables should be made using \code{numeric_variable()} or \code{nominal_variable()}
and passed directly as arguments.
}
\examples{

age_years <- numeric_variable(
  name = "age",
  label = "Age of participant",
  units = "years"
)

gender <- nominal_variable(
  name = "gender",
  label = "Gender of participant",
  category_levels = c("M", "F"),
  category_labels = c("Male", "Female")
)

dict <- data_dictionary(age_years, gender)
print(dict)

}
