% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDictionary.R
\name{date_variable}
\alias{date_variable}
\title{Create a Date Variable}
\usage{
date_variable(name, label = NULL, description = NULL, date_format = NULL)
}
\arguments{
\item{name}{Variable name}

\item{label}{Variable label}

\item{description}{Optional description}

\item{date_format}{Optional date format (e.g. "\%Y-\%m-\%d")}
}
\value{
A \code{DateVariable} object.
}
\description{
Create a Date Variable
}
