% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medication_coverage.R
\name{dt_add_coverage_dates}
\alias{dt_add_coverage_dates}
\title{todo - document}
\usage{
dt_add_coverage_dates(
  dt,
  max_carry = Inf,
  id_col = "id",
  med_col = "medication",
  fill_date_col = "date_fill",
  supply_col = "supply"
)
}
\arguments{
\item{dt}{todo}

\item{max_carry}{todo}

\item{id_col}{todo}

\item{med_col}{todo}

\item{fill_date_col}{todo}

\item{supply_col}{todo}
}
\value{
a \code{data.table}
}
\description{
todo - document
}
