% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_dictionary}
\alias{get_dictionary}
\title{Get data dictionary as a data frame}
\usage{
get_dictionary(x)
}
\arguments{
\item{x}{a \code{DataDictionary} object}
}
\value{
a \code{tibble} containing the data dictionary information
}
\description{
Get data dictionary as a data frame
}
\examples{

age_years <- numeric_variable(
  name = "age",
  label = "Age of participant"
)

age_group <- nominal_variable(
  name = "age_group",
  label = "Age group",
  category_levels = c("age_lt_50", "age_gteq_50_lt_60", "age_gteq_60"),
  category_labels = c("0 to < 50", "50 to < 60", "\u2265 60")
)

dd <- data_dictionary(age_years, age_group)

get_dictionary(dd)

}
