% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDictionary.R
\name{identifier_variable}
\alias{identifier_variable}
\title{Create an Identifier Variable}
\usage{
identifier_variable(name, label = NULL, description = NULL)
}
\arguments{
\item{name}{Variable name}

\item{label}{Variable label}

\item{description}{Optional description}
}
\value{
A \code{DateVariable} object.
}
\description{
Create an Identifier Variable
}
