% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infuse_dictionary.R
\name{infuse_dictionary}
\alias{infuse_dictionary}
\title{Infuse data with a dictionary}
\usage{
infuse_dictionary(
  data,
  dictionary = NULL,
  units = "none",
  warn_undocumented = TRUE
)
}
\arguments{
\item{data}{a data frame}

\item{dictionary}{A \link[=data_dictionary]{data dictionary} object.}

\item{units}{a character value indicating how units should be incorporated
into variable labels. Valid choices are
\itemize{
\item \code{"none"} : do not include units in labels, e.g., "Age"
\item \code{"descriptive"} : include labels, e.g., "Age, \emph{years}"
\item \code{"model"} : include label and model divisor, e.g., "Age, \emph{per 10 years}"
}}

\item{warn_undocumented}{a logical value. If \code{TRUE} (default), then
a warning is thrown whenever 1 or more variables in \code{data} do not
have supporting documentation in \code{dictionary}. If \code{FALSE}, then
this information will not be presented in a warning.}
}
\description{
Transfer variable and category labels into a dataframe. This
is useful if you are passing a data frame to a tabulation or plotting
function that naturally incorporates labeled data.
}
\details{
Additional details...
}
\examples{
age_years <- numeric_variable(
  name = "age_years",
  label = "Age of participant",
  units = "years",
  divby_modeling = 10
)

gender <- nominal_variable(
  name = "gender",
  label = "Gender of participant",
  category_levels = c("M", "F"),
  category_labels = c("Male", "Female")
)

dd <- data_dictionary(age_years, gender)
data <- data.frame(age_years = 55, gender = "M")

data \%>\%
  infuse_dictionary(dd, units = "model") \%>\%
  purrr::map(~attr(.x, 'label'))
}
