% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_factors}
\alias{set_factors}
\title{Set Factors
internal function used for factor modification}
\usage{
set_factors(dictionary, ..., .relevel)
}
\description{
Set Factors
internal function used for factor modification
}
