% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_each_group.R
\name{summarize_each_group}
\alias{summarize_each_group}
\title{Extends dplyr::summarize}
\usage{
summarize_each_group(data, ..., groups = NULL, include_overall = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{...}{R expressions to run in \code{summarize}}

\item{groups}{character vector with data names to use as groups}

\item{include_overall}{logical or character value. If \code{TRUE}, then
the summary will be computed in each group as well as overall, and
the overall group will be called ".overall". If \code{FALSE}, no overall
stat is included. If \code{include_overall} is provided as a character value,
e.g., \code{include_overall = "foo"}, then the summary will be computed
in each group as well as overall, and the overall group will be called
\code{"foo"}.}
}
\value{
a tibble
}
\description{
The extensions are:
}
\details{
\enumerate{
\item An overall result is always included.
\item Multiple groups can be summarized over.
}

either pass a grouped data frame or provide \code{groups} input,
and this function will run summarize based on what is provided
in \code{...} overall and in each group, separately.
}
\examples{

library(dplyr, quietly = TRUE)

summarize_each_group(iris,
                     groups = c("Species"),
                     across(c(matches("Sepal|Petal")), mean))


iris \%>\%
  group_by(Species) \%>\%
  summarize_each_group(n = n())
}
