% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pericircle-package.R
\docType{package}
\name{pericircle-package}
\alias{pericircle}
\alias{pericircle-package}
\title{pericircle: Easily Install and Load Perisphere Packages}
\description{
This package is designed to make it easy to install and load multiple packages used and developed by Perisphere Real World Evidence, LLC.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/perisphere-rwe/pericircle/issues}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@perisphere-rwe.com}

Authors:
\itemize{
  \item Tyler Sagendorf \email{tsagendorf@perisphere-rwe.com}
}

Other contributors:
\itemize{
  \item Perisphere Real World Evidence, LLC [copyright holder, funder]
}

}
\keyword{internal}
