% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pericircle-package.R
\docType{package}
\name{pericircle-package}
\alias{pericircle}
\alias{pericircle-package}
\title{pericircle: A Slice of Perisphere}
\description{
Frequently used tools in Perisphere's analyses designed to prevent redundancies and promote readable, maintainable code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcjaeger/pericircle}
  \item Report bugs at \url{https://github.com/bcjaeger/pericircle/issues}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bjaeger@wakehealth.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

}
\keyword{internal}
