% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_label}
\alias{set_label}
\alias{set_units}
\alias{set_description}
\alias{set_divby_modeling}
\alias{set_category_levels}
\alias{set_category_labels}
\title{Title}
\usage{
set_label(x, ...)

set_units(x, ...)

set_description(x, ...)

set_units(x, ...)

set_divby_modeling(x, ...)

set_category_levels(x, ...)

set_category_labels(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
